+++ src/demo/demo.c
@@ -170,6 +170,8 @@ static tb_demo_t g_demo[] =
 ,   TB_DEMO_MAIN_ITEM(platform_directory)
 ,   TB_DEMO_MAIN_ITEM(platform_cache_time)
 ,   TB_DEMO_MAIN_ITEM(platform_environment)
+,   TB_DEMO_MAIN_ITEM(platform_pipe_pair)
+,   TB_DEMO_MAIN_ITEM(platform_named_pipe)
 ,   TB_DEMO_MAIN_ITEM(platform_lock)
 ,   TB_DEMO_MAIN_ITEM(platform_timer)
 ,   TB_DEMO_MAIN_ITEM(platform_ltimer)