+++ src/tbox/platform/posix/file.c
@@ -91,6 +91,7 @@ tb_file_ref_t tb_file_init(tb_char_t const* path, tb_size_t mode)
             fd = open(path, flags, modes);
 #endif
     }
+    tb_check_return_val(fd >= 0, tb_null);
  
     // trace
     tb_trace_d("open: %p", tb_fd2file(fd));
@@ -104,7 +105,7 @@ tb_bool_t tb_file_exit(tb_file_ref_t file)
     tb_assert_and_check_return_val(file, tb_false);
 
     // trace
-    tb_trace_d("clos: %p", file);
+    tb_trace_d("close: %p", file);
 
     // close it
     tb_bool_t ok = !close(tb_file2fd(file))? tb_true : tb_false;