+++ src/tbox/platform/posix/socket_select.c
@@ -28,10 +28,6 @@
 #   include <sys/socket.h>
 #   include <sys/select.h>
 #endif
-#if defined(TB_CONFIG_MODULE_HAVE_COROUTINE) \
-        && !defined(TB_CONFIG_MICRO_ENABLE)
-#   include "../../coroutine/coroutine.h"
-#endif
 
 /* //////////////////////////////////////////////////////////////////////////////////////
  * macros
@@ -46,20 +42,10 @@
 /* //////////////////////////////////////////////////////////////////////////////////////
  * implementation
  */
-tb_long_t tb_socket_wait(tb_socket_ref_t sock, tb_size_t events, tb_long_t timeout)
+tb_long_t tb_socket_wait_impl(tb_socket_ref_t sock, tb_size_t events, tb_long_t timeout)
 {
     // check
     tb_assert_and_check_return_val(sock, -1);
-
-#if defined(TB_CONFIG_MODULE_HAVE_COROUTINE) \
-        && !defined(TB_CONFIG_MICRO_ENABLE)
-    // attempt to wait it in coroutine
-    if (tb_coroutine_self())
-    {
-        // wait it
-        return tb_coroutine_waitio(sock, events, timeout);
-    }
-#endif
 
     // fd
     tb_long_t fd = tb_sock2fd(sock);